backend/pyproject.toml
[project]
name = "call-analysis"
version = "0.1.0"
description = "AI-powered call center sentiment analysis and sale prediction system"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "librosa>=0.10.0",
    "openai-whisper>=20230314",
    "pyannote-audio>=3.3.2,<4.0.0",
    "scikit-learn>=1.3.0",
    "xgboost>=1.7.0",
    "pandas>=2.0.0",
    "numpy>=2.0.2,<2.3.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "plotly>=5.15.0",
    "flask>=2.3.0",          # Legacy Flask app (kept for reference)
    "flask-cors>=4.0.0",
    "fastapi>=0.115.0",      # New FastAPI backend
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.9",  # For FastAPI file uploads
    "pydub>=0.25.0",
    "soundfile>=0.12.0",
    "whisperx>=3.7.0",
    "resemblyzer>=0.1.1",
    "scipy>=1.10.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "pymongo>=4.15.1",
    "jiwer>=4.0.0",
    "spacy>=3.8.6",
    "download>=0.3.5",
    "reportlab",
    "thinc>=8.3.4,<8.4.0",
    "pip>=25.2",
    "psutil>=7.1.3",
    
]

[project.scripts]
call-analysis = "call_analysis:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/call_analysis"]
